+++ api/src/main/java/com/google/sps/ServiceCreation.java
@@ -1,48 +1,33 @@
 package com.google.sps;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.security.GeneralSecurityException;
 import java.util.Collections;
 import java.util.List;
-import java.util.Properties;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import org.springframework.util.ResourceUtils;
 
-import javax.mail.MessagingException;
-import javax.mail.Session;
-import javax.mail.internet.*;
-
 import com.google.api.client.json.JsonFactory;
-import com.google.api.client.auth.oauth2.BearerToken;
-import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
 import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
 import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
 import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
-import com.google.api.client.http.GenericUrl;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.jackson2.JacksonFactory;
 import com.google.api.client.util.store.FileDataStoreFactory;
 import com.google.api.services.gmail.Gmail;
 import com.google.api.services.gmail.GmailScopes;
-import com.google.api.services.gmail.model.Message;
-import com.google.auth.oauth2.GoogleCredentials;
 import com.google.api.client.http.HttpRequestInitializer;
-import com.google.auth.http.HttpCredentialsAdapter;
 
-import org.apache.commons.codec.binary.Base64;
-
-public class Notifications {
+public class ServiceCreation {
   private static final String APPLICATION_NAME = "Notifications";
   private static final JsonFactory JSON_FACTORY = JacksonFactory.getDefaultInstance();
   private static final String TOKENS_DIRECTORY_PATH = "tokens123";
@@ -83,56 +68,13 @@ public class Notifications {
     return new AuthorizationCodeInstalledApp(flow, receiver).authorize("user");
   }
 
-  private static MimeMessage createEmail(String to, String from, String subject, String bodyText)
-      throws MessagingException {
-    Properties props = new Properties();
-    Session session = Session.getDefaultInstance(props, null);
-
-    MimeMessage email = new MimeMessage(session);
-
-    email.setFrom(new InternetAddress(from));
-    email.addRecipient(javax.mail.Message.RecipientType.TO, new InternetAddress(to));
-    email.setSubject(subject);
-    email.setText(bodyText);
-    return email;
-  }
-
-  private static Message createMessageWithEmail(MimeMessage emailContent) throws MessagingException, IOException {
-    ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-    emailContent.writeTo(buffer);
-    byte[] bytes = buffer.toByteArray();
-    String encodedEmail = Base64.encodeBase64URLSafeString(bytes);
-    Message message = new Message();
-    message.setRaw(encodedEmail);
-    return message;
-  }
-
-  private static Message sendMessage(Gmail service, String userId, MimeMessage emailContent)
-    throws MessagingException, IOException {
-  Message message = createMessageWithEmail(emailContent);
-  message = service.users().messages().send(userId, message).execute();
-
-  System.out.println("Message id: " + message.getId());
-  System.out.println(message.toPrettyString());
-      return message;
-  }
-
-  public static void sendNotification(List<Email> emails) throws IOException, GeneralSecurityException {
+  public static Gmail createService() throws IOException, GeneralSecurityException {
     // Build a new authorized API client service.
     final NetHttpTransport HTTP_TRANSPORT = GoogleNetHttpTransport.newTrustedTransport();
     Gmail service = new Gmail.Builder(HTTP_TRANSPORT, JSON_FACTORY, getCredentials(HTTP_TRANSPORT))
             .setApplicationName(APPLICATION_NAME)
             .build();
-            
-    String from = "pictophone.noreply@gmail.com";
 
-    try {
-      for(Email email: emails) {
-        MimeMessage encoded = createEmail(email.player.getEmail(), from, email.getSubject(), email.getBody());
-        Message testMessage = sendMessage(service, from, encoded);
-      }
-    } catch(Exception e) {
-      System.out.println("Main Method Exception: " + e);
-    }
+    return service;
   }
 }
